ui/requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
filelock==3.17.0
fsspec==2025.2.0
imageio==2.37.0
Jinja2==3.1.5
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
scikit-image==0.25.2
scipy==1.15.2
setuptools==75.8.1
sympy==1.13.1
tifffile==2025.2.18
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
